event states update, with command and controllers

+++ src/Controller/SubscriptionController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Event;
 use App\Entity\EventState;
 use App\Entity\EventSubscription;
+use App\EventState\EventStateHelper;
 use App\Form\EventType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -16,27 +17,32 @@ class SubscriptionController extends AbstractController
     /**
      * @Route("/sorties/{id}/inscription/", name="subscription_toggle")
      */
-    public function toggle(Event $event)
+    public function toggle(Event $event, EventStateHelper $stateHelper)
     {
         $em = $this->getDoctrine()->getManager();
-
         $subscriptionRepo = $this->getDoctrine()->getRepository(EventSubscription::class);
-        $foundSubscription = $subscriptionRepo->findOneBy(['user' => $this->getUser(), 'event' => $event]);
+
+        //un peu de validations
+        if ($event->getState()->getName() !== "open"){
+            $this->addFlash("danger", "Cette sortie n'est pas ouverte aux inscriptions !");
+            return $this->redirectToRoute('event_detail', ["id" => $event->getId()]);
+        }
 
         //désincription si on trouve cette inscription
+        $foundSubscription = $subscriptionRepo->findOneBy(['user' => $this->getUser(), 'event' => $event]);
         if ($foundSubscription){
             $em->remove($foundSubscription);
             $em->flush();
 
             $this->addFlash("success", "Vous êtes désinscrit !");
-            return $this->redirectToRoute('event_list');
+            return $this->redirectToRoute('event_detail', ["id" => $event->getId()]);
         }
 
         //sinon, inscription
         //complet ?
-        if ($event->getMaxRegistrations() !== null && $event->getSubscriptions()->count() >= $event->getMaxRegistrations()){
+        if ($event->isMaxedOut()){
             $this->addFlash("danger", "Cette sortie est complète !");
-            return $this->redirectToRoute('event_list');
+            return $this->redirectToRoute('event_detail', ["id" => $event->getId()]);
         }
 
         $subscription = new EventSubscription();
@@ -46,7 +52,14 @@ class SubscriptionController extends AbstractController
         $em->persist($subscription);
         $em->flush();
 
+        $em->refresh($event);
+
+        //maintenant, on tchèque si c'est complet pour changer son état
+        if ($event->isMaxedOut()){
+            $stateHelper->changeEventState($event, 'closed');
+        }
+
         $this->addFlash("success", "Vous êtes inscrit !");
-        return $this->redirectToRoute('event_list');
+        return $this->redirectToRoute('event_detail', ["id" => $event->getId()]);
     }
 }
